Value graph in double precision
Scene time was already double (FrameContext::seconds) but truncated to
f32 the moment it entered the value graph, so any time-driven animation
turned steppy after ~a day of uptime (f32 ULP at 30 days is 0.25s —
larger than a frame step). Value now carries f64 components end to end;
narrowing to f32 happens only where uniforms are packed for the GPU
(transform and shader nodes). wave additionally reduces its phase to
one cycle in double before trig so the argument stays bounded at any
uptime, and JSON literals no longer round-trip through float at parse.

Unit tests drive time->wave at a month and a year of scene time; both
would fail with an f32 value path.

+++ src/core/Scene.cpp
@@ -15,7 +15,7 @@ Value Node::inputValue(size_t index) const
         v = in.constant;
     }
     if (in.splat && v.type == ValueType::Scalar && in.type != ValueType::Scalar) {
-        const float s = v.v[0];
+        const double s = v.v[0];
         v.type = in.type;
         v.v = { s, s, s, s };
     }
@@ -46,7 +46,7 @@ bool Scene::setParameter(const std::string& name, Value value)
             continue;
         }
         if (value.type == ValueType::Scalar && param.type != ValueType::Scalar) {
-            const float s = value.v[0];
+            const double s = value.v[0];
             value.type = param.type;
             value.v = { s, s, s, s };
         }
